Feature/extract download to dataset download part2 (#187)
* 3.1 complete

* Refactor download dialog helpers to new location

- Moved download dialog helper functions to '../../../features/datasetDownload/utils/downloadDialogHelpers'.
- Updated import paths in DateSubsetControl.js, DialogContent.js, and downloadSagas.js to reflect the new location.
- Removed the old downloadDialogHelpers.js file as it is no longer needed.

* Remove legacy download dialog components and update import paths

- Deleted buttonStates.js and downloadDialogStyles.js as they are no longer needed.
- Updated import paths in various components to reflect the new location of styles and button states.
- Ensured all references to removed files are updated to maintain functionality.

* Refactor download dialog components by removing legacy files and updating imports

- Deleted ErrorMessage.js and ValidationIndicatorBar.js as they are no longer needed.
- Updated import paths in DialogContent.js to reflect the new location of ErrorMessage and ValidationIndicatorBar.
- Ensured all references to removed files are updated to maintain functionality.

* Refactor download dialog components by removing Header.js and updating import paths

- Deleted Header.js as it is no longer needed.
- Updated import path in DialogContent.js to reflect the new location of DownloadDialogTitle.
- Ensured all references to removed files are updated to maintain functionality.

* finish intro

* Remove StatusBar component and update import paths in ValidationIndicatorBar.js

* b2 complwtw

* Refactor download dialog components by removing unused files and updating import paths

- Deleted DownloadStep.js and DownloadStepWithWarning.js as they are no longer needed.
- Updated import paths in DialogContent.js to reflect the new locations of DownloadStep and DownloadStepWithWarning.
- Ensured all references to removed files are updated to maintain functionality.

* wip

* wip remaining

* wip

---------
