+++ src/features/datasetDownload/components/DownloadButtons/index.js
@@ -1,0 +1,6 @@
+// DownloadButtons component exports
+// Files will be moved here: DownloadButtons.js
+
+export { default as DownloadButtons } from './DownloadButtons';
+export { default as DownloadButtonOutlined } from './DownloadButtons';
+export { default as DownloadButtonFilled } from './DownloadButtons';
