+++ src/features/datasetDownload/components/DownloadDialog/Header.js
@@ -1,7 +1,7 @@
 import { DialogTitle, Typography } from '@material-ui/core';
 import { withStyles } from '@material-ui/core/styles';
 import React from 'react';
-import styles from './downloadDialogStyles';
+import styles from '../../styles/downloadDialogStyles';
 
 const Title = (props) => {
   let { longName, classes } = props;
