+++ src/features/datasetDownload/components/DownloadDialog/StatusBar.js
@@ -15,7 +15,7 @@ import WarningIcon from '@material-ui/icons/Warning';
 import CheckCircleIcon from '@material-ui/icons/CheckCircle';
 import NotInterestedIcon from '@material-ui/icons/NotInterested';
 
-import { buttonStates } from './buttonStates';
+import { buttonStates } from '../../utils/buttonStates';
 
 const useStyles = makeStyles((theme) => ({
   root: {
