+++ src/features/datasetDownload/components/DownloadDialog/index.js
@@ -1,0 +1,9 @@
+// DownloadDialog component exports
+// Files will be moved here: DialogContent.js, GlobalDialogWrapper.js, Header.js, Intro.js, StatusBar.js, DownloadStep.js
+
+export { default as DialogContent } from './DialogContent';
+export { default as GlobalDialogWrapper } from './GlobalDialogWrapper';
+export { default as Header } from './Header';
+export { default as Intro } from './Intro';
+export { default as StatusBar } from './StatusBar';
+export { default as DownloadStep } from './DownloadStep';
