+++ src/features/datasetDownload/components/DropboxIntegration/index.js
@@ -1,0 +1,5 @@
+// DropboxIntegration component exports
+// Files will be moved here: DropboxModal.js, DropboxEmbed.js
+
+export { default as DropboxModal } from './DropboxModal';
+export { default as DropboxEmbed } from './DropboxEmbed';
