+++ src/features/datasetDownload/components/Helpers/index.js
@@ -1,0 +1,7 @@
+// Helper component exports
+// Files will be moved here: ValidationIndicatorBar.js, EmbeddedLogin.js, ErrorMessage.js, AncillaryDataDownload.js
+
+export { default as ValidationIndicatorBar } from './ValidationIndicatorBar';
+export { default as EmbeddedLogin } from './EmbeddedLogin';
+export { default as ErrorMessage } from './ErrorMessage';
+export { default as AncillaryDataDownload } from './AncillaryDataDownload';
