+++ src/features/datasetDownload/components/SubsetControls/DateSubsetControl.js
@@ -12,8 +12,8 @@ import {
   dayToDateString,
   dateToDateString,
   getIsMonthlyClimatology,
-} from './downloadDialogHelpers';
-import styles from './downloadDialogStyles';
+} from '../../utils/downloadDialogHelpers';
+import styles from '../../styles/downloadDialogStyles';
 
 // convert a date string like "2007-04-09" to "4/9"
 let shortenDate = (str) =>
