+++ src/features/datasetDownload/components/SubsetControls/DepthSubsetControl.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Grid, Slider, TextField, Typography } from '@material-ui/core';
 import { withStyles } from '@material-ui/core/styles';
-import styles from './downloadDialogStyles';
+import styles from '../../styles/downloadDialogStyles';
 
 const DepthSubsetControl = (props) => {
   let { classes, dataset, subsetState, setDepthStart, setDepthEnd } = props;
