+++ src/features/datasetDownload/components/SubsetControls/LongitudeSubsetControl.js
@@ -1,11 +1,11 @@
 import { Grid, Slider, TextField, Typography } from '@material-ui/core';
 import { withStyles } from '@material-ui/core/styles';
 import React from 'react';
-import styles from './downloadDialogStyles';
-import logInit from '../../../Services/log-service';
+import styles from '../../styles/downloadDialogStyles';
+import logInit from '../../../../Services/log-service';
 
 const log = logInit('LongitudeSubsetControl').addContext({
-  src: 'Components/Catalog/DownloadDialog',
+  src: 'features/datasetDownload/components/DownloadDialog',
 });
 
 const LongitudeControl = (props) => {
