+++ src/features/datasetDownload/components/SubsetControls/SubsetControls.js
@@ -5,11 +5,11 @@ import DateSubsetControl from './DateSubsetControl';
 import LatitudeSubsetControl from './LatitudeSubsetControl';
 import LongitudeSubsetControl from './LongitudeSubsetControl';
 import DepthSubsetControl from './DepthSubsetControl';
-import DownloadOption from './DownloadOption';
+import DownloadOption from '../DownloadDialog/DownloadOption';
 
-import logInit from '../../../Services/log-service';
+import logInit from '../../../../Services/log-service';
 const log = logInit('dialog subset controls').addContext({
-  src: 'Components/Catalog/DownloadDialog',
+  src: 'features/datasetDownload/components/SubsetControls',
 });
 
 // This component gathers the four sets of control fields and passes
