+++ src/features/datasetDownload/components/SubsetControls/index.js
@@ -1,0 +1,8 @@
+// SubsetControls component exports
+// Files will be moved here: SubsetControls.js, DateSubsetControl.js, DepthSubsetControl.js, LatitudeSubsetControl.js, LongitudeSubsetControl.js
+
+export { default as SubsetControls } from './SubsetControls';
+export { default as DateSubsetControl } from './DateSubsetControl';
+export { default as DepthSubsetControl } from './DepthSubsetControl';
+export { default as LatitudeSubsetControl } from './LatitudeSubsetControl';
+export { default as LongitudeSubsetControl } from './LongitudeSubsetControl';
