+++ src/features/datasetDownload/components/index.js
@@ -1,4 +1,16 @@
-// Dataset Download Component exports
-// Placeholder for download dialog and related components
+// Main components index - exports from all component subdirectories
 
-export {};
+// DownloadDialog components
+export * from './DownloadDialog';
+
+// DownloadButtons components
+export * from './DownloadButtons';
+
+// SubsetControls components
+export * from './SubsetControls';
+
+// DropboxIntegration components
+export * from './DropboxIntegration';
+
+// Helper components
+export * from './Helpers';
