+++ src/features/datasetDownload/utils/index.js
@@ -1,4 +1,5 @@
-// Dataset Download Utility exports
-// Placeholder for download validation, query builders, and helpers
+// Utility exports
+// Files will be moved here: downloadDialogHelpers.js, buttonStates.js
 
-export {};
+export * from './downloadDialogHelpers';
+export * from './buttonStates';
